requirements.txt
Python version 3.9.6
pip version 21.3.1
tensorflow version 2.7.0
pandas version 1.3.4
tensorflow_datasets-4.4.0
sklearn (will install joblib-1.1.0 scikit-learn-1.0.1 scipy-1.7.3 sklearn-0.0 threadpoolctl-3.0.0)
nltk 3.6.5
matplotlib 3.5.0
transformers 4.12.5
cuda_11.5.1_496.13_windows
Download glove 6B 50d, 100d, 200d and 300d and put it in the path .\SentimentAnalysis-ComparitiveStudy-FlaskAPI-Predict\data\glove.6B
BERT needs to be trained, which will create a folder BERTsentiment which will contain config.json and tf_model.h5.